double uninstall errors fixed, target print added

+++ Makefile
@@ -25,45 +25,66 @@ help:
 all: help
 usage: help
 
-install: test conf var pkg masterconf installcomplete
+install: test conf pkg var masterconf installcomplete
+
+installapp: test conf-app pkg-app var masterconf installcomplete
 
 conf:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; [ -d conf ] || mkdir conf ; root=$$PWD ; popd > /dev/null ; \
 pushd ws/eg > /dev/null ; \
 for p in conf/*.{json,conf} ; do [ -f $$root/$$p ] || ln -s  ${PWD}/ws/eg/$$p $$root/$$p ; done ; \
 popd > /dev/null
 
+conf-app:
+	@echo "*** $@ ***"
+	pushd .. > /dev/null ; [ -d conf ] || mkdir conf ; root=$$PWD ; \
+pushd pkg/app/eg > /dev/null ; \
+for p in conf/*.json ; do [ -f $$root/$$p ] || ln -s $$root/pkg/app/eg/$$p $$root/$$p ; done ; \
+popd > /dev/null ; popd > /dev/null
+
 pkg:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; [ -d pkg ] || mkdir pkg ; root=$$PWD ; popd > /dev/null ; \
 pushd ws/eg > /dev/null ; \
 for p in pkg/* ; do [ -e $$root/$$p ] || ln -s $$PWD/$$p $$root/$$p ; done ; \
 popd > /dev/null
 
+pkg-app:
+	@echo "*** $@ ***"
+	pushd .. > /dev/null ; root=$$PWD ; . pgws.cfg ; \
+for p in $$PGWS_APP_PKG ; do [ -e $$root/pkg/$$p ] || ln -s $$root/pgws/ws/eg/pkg/$$p pkg/$$p ; done ; \
+popd > /dev/null
+
 var:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 [ -d var ] || mkdir var ; \
 for p in build cache log run tmpl tmpc ; do [ -d var/$$p ] || mkdir -m g+w var/$$p ; done ; \
 popd > /dev/null
 
 masterconf:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 [ -f pgws.cfg ] || ln ${PWD}/ws/eg/pgws.cfg pgws.cfg ; \
 [ -f pgws.sh ] || ln ${PWD}/ws/eg/pgws.sh pgws.sh ; \
 popd > /dev/null
 
 gitignore:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 [ -f .gitignore ] || touch .gitignore ; \
 for p in pgws var conf  ; do grep -E "^$$p" .gitignore > /dev/null || echo "$$p/*" >> .gitignore ; done ; \
 popd > /dev/null
 
 installcomplete:
-	@echo "Install complete."
+	@echo "*** $@ ***"
 
 uninstallcomplete:
-	@echo "UnInstall complete."
+	@echo "*** $@ ***"
 
 clean:
+	@echo "*** $@ ***"
 	rm -rf .install .usage ; \
 pushd .. > /dev/null ; \
 [ -f var/run/*.pid ] && ./pgws.sh fcgi stop ; \
@@ -71,33 +92,38 @@ pushd .. > /dev/null ; \
 popd > /dev/null
 
 uninstall-db:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 [ -f var/.build.pkg ] && ./pgws.sh db drop pkg ; \
 [ -f var/.build.pgws ] && ./pgws.sh db drop ; \
 popd > /dev/null
 
 uninstall-dirs:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 [ -d var ] && for p in var/{build,cache,log,run,tmpl,tmpc} ; do rmdir $$p 2> /dev/null ; done ; \
 for p in var pkg conf ; do rmdir $$p 2> /dev/null ; done ; \
 popd > /dev/null
 
 clean-conf:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 for p in conf/*.{json,conf} ; do [ $$p -ef pgws/ws/eg/$$p ] && rm $$p ; done ; \
-for f in pgws.sh pgws.cfg ; do diff --brief $$f pgws/ws/eg/$$f && rm $$f ; done ; \
+for f in pgws.sh pgws.cfg ; do [ -f $$f ] && diff --brief $$f pgws/ws/eg/$$f && rm $$f ; done ; \
 popd > /dev/null
 
 clean-pkg:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 [ -s var/i18n ] && [ var/i18n -ef pkg/i18n/src/templates ] && rm var/i18n ; \
-for p in pkg/* ; do [ $$p -ef pgws/ws/eg/$$p ] && rm $$p ; done ; \
+[ -d pkg ] && for p in pkg/* ; do [ $$p -ef pgws/ws/eg/$$p ] && rm $$p ; done ; \
 popd > /dev/null
 
 clean-lib:
+	@echo "*** $@ ***"
 	pushd .. > /dev/null ; \
 R=$$(dirs +0) ; \
-find lib -name *.pm -printf "%p %l\n" | while read p s ; do \
+[ -d lib ] && find lib -name *.pm -printf "%p %l\n" | while read p s ; do \
   if [[ "$$s" && "$$s" != "$${s#$$R/pgws/ws/lib/}" ]] ; then \
     rm $$p ; \
   else \
@@ -110,7 +136,7 @@ find lib -name *.pm -printf "%p %l\n" | while read p s ; do \
     fi ; \
   fi ; \
 done ; \
-find lib -type d | while read p ; do rmdir -p $$p 2>> /dev/null ; done ; \
+[ -d lib ] && find lib -type d | while read p ; do rmdir -p $$p 2>> /dev/null ; done ; \
 popd > /dev/null
 
 uninstall: uninstall-db clean clean-lib clean-pkg clean-conf uninstall-dirs uninstallcomplete
